refactor for reporters directory
